Rename namespace for IServiceMessageHandler
