+++ source/Sashimi.Aws.Accounts/AmazonWebServicesAccountModule.cs
@@ -1,6 +1,7 @@
 using Autofac;
 using Octopus.Server.Extensibility.Extensions.Mappings;
 using Sashimi.Server.Contracts.Accounts;
+using Sashimi.Server.Contracts.ServiceMessages;
 
 namespace Sashimi.Aws.Accounts
 {
